Move code bulks to separate files
