+++ src/args.ts
@@ -15,32 +15,6 @@ export interface Args {
 }
 
 /**
- * Help message to be shown if `denon`
- * is run with `--help` flag.
- */
-export function help(version: string): string {
-  setColorEnabled(true);
-  return `
-${blue("DENON")} - ${version}
-Monitor any changes in your Deno application and automatically restart.
-
-Usage:
-    ${blue("denon")} ${yellow("<script name>")}     ${
-    gray("-- eg: denon start")
-  }
-    ${blue("denon")} ${yellow("<command>")}         ${
-    gray("-- eg: denon run helloworld.ts")
-  }
-    ${blue("denon")} [options]         ${gray("-- eg: denon --help")}
-
-Options:
-    -h --help          Show this screen.
-    -v --version       Show version.
-    -i --init          Create config file in current working dir.
-`;
-}
-
-/**
  * Parse Deno.args into a flag map (`Args`) 
  * to be handled by th CLI.
  */
